chore(CustomerIO): fix schema
